Compare and align daily progress

+++ Hub/Hub.Application/Models/Progress/PlayerProgressGetModel.cs
@@ -5,5 +5,5 @@ namespace Hub.Application.Models.Progress;
 public class PlayerProgressGetModel
 {
     [JsonPropertyName("progress")]
-    public Dictionary<string, decimal>? Progress { get; set; }
+    public Dictionary<string, int>? Progress { get; set; }
 }
